Add end to end working solution for in memory index using macros

+++ src/sharded_memmap_index.rs
@@ -61,7 +61,7 @@ impl ShardedMemmapIndex {
     }
 
     #[pyo3(signature = (query, vocab=None))]
-    pub fn count_next(&self, query: Vec<u16>, vocab: Option<u16>) -> Vec<usize> {
+    pub fn count_next(&self, query: Vec<u16>, vocab: Option<usize>) -> Vec<usize> {
         let counts = self
             .shards
             .iter()
@@ -73,7 +73,7 @@ impl ShardedMemmapIndex {
     }
 
     #[pyo3(signature = (queries, vocab=None))]
-    pub fn batch_count_next(&self, queries: Vec<Vec<u16>>, vocab: Option<u16>) -> Vec<Vec<usize>> {
+    pub fn batch_count_next(&self, queries: Vec<Vec<u16>>, vocab: Option<usize>) -> Vec<Vec<usize>> {
         let batch_counts = self
             .shards
             .iter()
@@ -97,7 +97,7 @@ impl ShardedMemmapIndex {
         n: usize,
         k: usize,
         num_samples: usize,
-        vocab: Option<u16>,
+        vocab: Option<usize>,
     ) -> Result<Vec<Vec<u16>>> {
         self.sample_unsmoothed_rs(&query, n, k, num_samples, vocab)
     }
@@ -105,7 +105,7 @@ impl ShardedMemmapIndex {
     /// Returns interpolated Kneser-Ney smoothed token probability distribution using all previous
     /// tokens in the query.
     #[pyo3(signature = (query, vocab=None))]
-    pub fn get_smoothed_probs(&mut self, query: Vec<u16>, vocab: Option<u16>) -> Vec<f64> {
+    pub fn get_smoothed_probs(&mut self, query: Vec<u16>, vocab: Option<usize>) -> Vec<f64> {
         self.get_smoothed_probs_rs(&query, vocab)
     }
 
@@ -115,7 +115,7 @@ impl ShardedMemmapIndex {
     pub fn batch_get_smoothed_probs(
         &mut self,
         queries: Vec<Vec<u16>>,
-        vocab: Option<u16>,
+        vocab: Option<usize>,
     ) -> Vec<Vec<f64>> {
         self.batch_get_smoothed_probs_rs(&queries, vocab)
     }
@@ -128,7 +128,7 @@ impl ShardedMemmapIndex {
         n: usize,
         k: usize,
         num_samples: usize,
-        vocab: Option<u16>,
+        vocab: Option<usize>,
     ) -> Result<Vec<Vec<u16>>> {
         self.sample_smoothed_rs(&query, n, k, num_samples, vocab)
     }
@@ -151,7 +151,7 @@ impl Sample for ShardedMemmapIndex {
         &mut self.cache
     }
 
-    fn count_next_slice(&self, query: &[u16], vocab: Option<u16>) -> Vec<usize> {
+    fn count_next_slice(&self, query: &[u16], vocab: Option<usize>) -> Vec<usize> {
         let counts = self
             .shards
             .iter()
